# config/api-doc.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Alif's Backend App API Documentation",
        "version": "0.1.0",
        "description": "If You run into any problem feel free to contact me:",
        "contact": {
            "name": "Alif",
            "email": "nuralif.2020@student.uny.ac.id",
            "phone": "+6281217304001"
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
...
            "moviesResponse": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "description": "Movie title"
                    },
                    "poster": {
                        "type": "string",
                        "description": "poster url"
                    }
                },
                "example": {
                    "title": "Interstellar",
                    "poster": "https://www.image.jpg"
                }
            },
...
            "moviesUpdateFavorite": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "description": "New Movie title"
                    },
                    "posterUrl": {
                        "type": "string",
                        "description": "New URL"
                    }
                },
                "example": {
                    "title": "New Title",
                    "posterUrl": "New poster url"
                }
            },
...
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000/"
        }
    ],
    "paths": {
...
        "/movies/{title}": {
            "get": {
                "summary": "Return poster url of that movie",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "Movies"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "title",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "Movie Title"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "poster url",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/moviesResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Movie not found."
                    }
                }
            },
            "put": {
                "summary": "update poster url and movie name of that movie",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "Movies"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "title",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "Movie Title"
                    }
                ],"requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/moviesUpdateFavorite"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "moviesUpdated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/moviesResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Movie not found."
                    }
                }
            },
            "delete": {
                "summary": "Deletes a Favorite Movie by title",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "Movies"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "title",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "Movie Title"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Delete was successful."
                    },
                    "404": {
                        "description": "Favorite Movie not found."
                    }
                }
            }
        },
...
    },
...
}